+++ src/identity/punish/mod.rs
@@ -10,6 +10,9 @@ use crate::identity::{
     vouch::vouchees,
 };
 
+pub mod db;
+pub mod storage;
+
 // allows to ban for twice the entire balance, i.e. permanent ban. However due to penalty decay
 // IDT balance can eventually become positive.
 // It only limits vouchee penalty because we do not want to limit amount of penalties and their value
