+++ src/verify/admin.rs
@@ -1,7 +1,10 @@
 use crate::{
     identity::UserAddress,
     verify::{
-        Nonce, error::Error, nonce::NonceManager, private_key_to_address, signature::Signature,
+        error::Error,
+        nonce::{Nonce, NonceManager},
+        private_key_to_address,
+        signature::Signature,
     },
 };
 
@@ -98,6 +101,6 @@ mod tests {
         let err = admin_verify(&signature, user, &nonce_manager)
             .await
             .unwrap_err();
-        assert!(matches!(err, Error::NonceUsedError(_)));
+        assert!(matches!(err, Error::NonceError(_)));
     }
 }
