+++ src/verify/mod.rs
@@ -8,13 +8,10 @@ pub mod error;
 pub mod forget;
 pub mod moderator;
 pub mod nonce;
-pub mod nonce_db;
 pub mod proof;
 pub mod punish;
 pub mod signature;
 pub mod vouch;
-
-pub type Nonce = u64;
 
 pub fn address_to_string(address: &H160) -> String {
     format!("0x{}", hex::encode(address.as_bytes()))
